gitlab_return_repositories_with_older_commits/requirements.txt
python-gitlab==5.6.0
python-dateutil==2.9.0
urllib3==2.3.0
